% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ssbmsy.R
\name{assign_ssbmsy}
\alias{assign_ssbmsy}
\title{Find Baseline SSBmsy}
\usage{
assign_ssbmsy(fauna, sel_start = 0.01, years = 50)
}
\arguments{
\item{fauna}{a fauna object}

\item{sel_start}{selectivity as a multiplier of length at maturity}

\item{years}{the number of years to run things out for}

\item{mult}{multiplier to effort}

\item{use}{one of "graphs" or "optim"}
}
\value{
results of simulation
}
\description{
And assigns it to each critter in fauna object
}
\examples{
\dontrun{
years <- 50

seasons <- 4

fauna <-
  list(
    "bigeye" = create_critter(
      scientific_name = "thunnus obesus",
      adult_diffusion = 10,
      density_dependence = "post_dispersal",
      seasons = seasons,
      resolution = resolution,
      age_mature = 1,
      steepness = 0.9,
      ssb0 = 1000
    )
  )

baseline_ssbmsy <- find_ssbmsy(fauna = fauna)
}

}
