% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_step.R
\name{clean_step}
\alias{clean_step}
\title{A small function to trim any characters from step names in the output of \code{marlin::simmar}}
\usage{
clean_step(step)
}
\arguments{
\item{step}{The name of the step in question}
}
\value{
a numeric value with only the numeric (plus decimal) parts of the step name
}
\description{
A small function to trim any characters from step names in the output of \code{marlin::simmar}
}
\examples{
clean_step("step_1")

}
