% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish-class.R
\name{Fish}
\alias{Fish}
\title{R6 Class Representing a Fish}
\description{
A fish object has all the life history data needed to model a finfish population
}
\details{
Fish like to swim
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Fish$new()}}
\item \href{#method-plot}{\code{Fish$plot()}}
\item \href{#method-clone}{\code{Fish$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
fill in fish object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fish$new(
  common_name = "white seabass",
  scientific_name = NA,
  linf = NA,
  vbk = NA,
  t0 = -0.1,
  cv_len = 0.1,
  length_units = "cm",
  min_age = 0,
  max_age = NA,
  weight_a = NA,
  weight_b = NA,
  weight_units = "kg",
  length_50_mature = NA,
  length_95_mature = NA,
  delta_mature = 0.1,
  age_50_mature = NA,
  age_95_mature = NA,
  age_mature = NA,
  length_mature = NA,
  m = NA,
  steepness = 0.8,
  r0 = 10000,
  ssb0 = NA,
  density_dependence_form = 1,
  adult_movement = 2,
  adult_movement_sigma = 2,
  larval_movement = 2,
  query_fishlife = T,
  sigma_r = 0,
  rec_ac = 0,
  cores = 4,
  mat_mode = "age",
  default_wb = 2.8,
  tune_weight = FALSE,
  density_movement_modifier = 1,
  linf_buffer = 1.2,
  resolution = 25,
  seasonal_habitat = list(),
  habitat_seasons = list(seasons),
  rec_habitat = NA,
  fished_depletion = 1,
  rec_form = 1,
  burn_years = 100,
  seasonal_hab = NA,
  seasons = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{common_name}}{common name}

\item{\code{seasons}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fish$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fish$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
