% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish-class-reserve.R
\name{Fishold}
\alias{Fishold}
\title{R6 Class Representing a Fish}
\description{
A fish object has all the life history data needed to model a finfish population
}
\details{
Fish like to swim
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Fishold-new}{\code{Fishold$new()}}
\item \href{#method-Fishold-plot}{\code{Fishold$plot()}}
\item \href{#method-Fishold-swim}{\code{Fishold$swim()}}
\item \href{#method-Fishold-clone}{\code{Fishold$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fishold-new"></a>}}
\if{latex}{\out{\hypertarget{method-Fishold-new}{}}}
\subsection{Method \code{new()}}{
fill in fish object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fishold$new(
  common_name = NA,
  scientific_name = NA,
  linf = NA,
  vbk = NA,
  t0 = -0.1,
  cv_len = 0.1,
  length_units = "cm",
  min_age = 0,
  max_age = NA,
  weight_a = NA,
  weight_b = NA,
  pups = 10,
  weight_units = "kg",
  fec_form = "weight",
  fec_expo = 1,
  length_50_mature = NA,
  length_95_mature = NA,
  delta_mature = 0.1,
  age_50_mature = NA,
  age_95_mature = NA,
  age_mature = NA,
  length_mature = NA,
  m = NA,
  steepness = 0.8,
  r0 = 10000,
  ssb0 = NA,
  density_dependence_form = 1,
  adult_diffusion = 4,
  recruit_diffusion = 10,
  query_fishlife = T,
  sigma_r = 0,
  rec_ac = 0,
  cores = 4,
  mat_mode = "age",
  default_wb = 2.8,
  tune_weight = FALSE,
  density_movement_modifier = 1,
  linf_buffer = 1.2,
  resolution = NA,
  seasonal_habitat = list(),
  season_blocks = list(),
  recruit_habitat = NA,
  fished_depletion = 1,
  rec_form = 1,
  burn_years = 50,
  seasonal_hab = NA,
  seasons = 1,
  explt_type = "f",
  init_explt = 0.1,
  get_common_name = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{common_name}}{common name}

\item{\code{pups}}{}

\item{\code{fec_form}}{}

\item{\code{fec_expo}}{exponent for fecundity at weight relationship, 1 = isometric > 1 hyperallometric}

\item{\code{explt_type}}{Swim

Swim advances the population one time step}

\item{\code{get_common_name}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fishold-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Fishold-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fishold$plot(type = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fishold-swim"></a>}}
\if{latex}{\out{\hypertarget{method-Fishold-swim}{}}}
\subsection{Method \code{swim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fishold$swim(
  burn_steps = 0,
  season = 1,
  f_p_a = NULL,
  last_n_p_a = NULL,
  adult_movement = NULL,
  tune_unfished = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{burn_steps}}{number of steps for burn in period if applicable}

\item{\code{season}}{the current season}

\item{\code{f_p_a}}{matrix of fishing mortality by patch and age}

\item{\code{last_n_p_a}}{matrix of initial numbers by patch and age}

\item{\code{tune_unfished}}{boolean indicating whether to tune unfished}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the population in the next time step
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fishold-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Fishold-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fishold$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
