% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_tuner.R
\name{fleet_tuner}
\alias{fleet_tuner}
\title{fleet tuner}
\usage{
fleet_tuner(log_fs, fauna, fleets, e_fl, years = 50)
}
\arguments{
\item{log_fs}{log instantaneous fishing mortality per critter}

\item{fauna}{fauna object}

\item{fleets}{fleet object}

\item{e_fl}{baseline effort per fleet}

\item{years}{number of years to tune}
}
\value{
objective function of fleet tuner
}
\description{
finds catchability (q) by fleet such that
target fished depletion is achieved
}
