% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mpa_distances.R
\name{get_distance_to_mpas}
\alias{get_distance_to_mpas}
\title{Measures the distance of each cell to the nearest MPA edge and to all MPA cells}
\usage{
get_distance_to_mpas(mpa_locations, resolution, patch_area = 10)
}
\arguments{
\item{mpa_locations}{a dataframe with at least column coordinate columns x,y,and mpa (TRUE or FALSE)}

\item{resolution}{the resolution of the simulation system. Can supply a vector \code{c(x,y)} to denote an X by Y system, or one number to denote an X by X system}

\item{patch_area}{the area of each patch}
}
\value{
a dataframe with two columns added.
\code{distance_to_mpa_edge} measures the distance from each patch to the nearest MPA edge in units of sqrt(patch_area), with negative values indicating areas inside an MPA
\code{total_mpa_distance} measure to total distance to all MPA cells from each patch, in units of sqrt(patch_area)
}
\description{
Measures the distance of each cell to the nearest MPA edge and to all MPA cells
}
