% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_fish.R
\name{go_fish}
\alias{go_fish}
\title{Go Fishing applies a given matrix of effort by patch and fleet to determine the
returns from fishing of that effort. This does not actually affect the population, this is
essentially an exploratory fishing step used to allocate fishing effort in space.}
\usage{
go_fish(e_p_fl, fauna, fleets, groupers = c("fleet", "patch"))
}
\arguments{
\item{e_p_fl}{a matrix of effort by patch and fleet}

\item{fauna}{a list of fauna objects}

\item{fleets}{a list of fleet objects}

\item{groupers}{the resolution to return the ouputs at}
}
\value{
a list of the different types of returns
}
\description{
Go Fishing applies a given matrix of effort by patch and fleet to determine the
returns from fishing of that effort. This does not actually affect the population, this is
essentially an exploratory fishing step used to allocate fishing effort in space.
}
