% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_mpa.R
\name{place_mpa}
\alias{place_mpa}
\title{Place MPA}
\usage{
place_mpa(
  target_fauna,
  size = 0.2,
  fauna,
  placement_error = 0,
  seed = 42,
  place_randomly = FALSE
)
}
\arguments{
\item{target_fauna}{a vector of critter names around which to base MPA}

\item{size}{the proportion of available patches to place in an MPA}

\item{fauna}{a fauna object}

\item{placement_error}{CV of log-normal error to apply to MPA citing}

\item{seed}{seed for random MPA stuff}

\item{place_randomly}{TRUE to place MPAs randomly weighted by something}
}
\value{
a data frame with coordinates of MPA locations
}
\description{
Locates MPAs at the moment based on numbers of arbitrary groups of species
}
