% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_recruitment_deviations.R
\name{sim_rec_devs}
\alias{sim_rec_devs}
\title{Simulate species habitats with correlations across space and among species}
\usage{
sim_rec_devs(
  critters,
  kp,
  critter_correlations = NA,
  resolution,
  patch_area,
  max_delta = 3,
  max_abs_cor = 1,
  output = "df"
)
}
\arguments{
\item{critters}{a vector of critters names of length n_species representing the number of species to be simulated}

\item{kp}{a rate parameter that governs the rate at which spatial cells become decorrelated with distance. Lower values mean a smoother (more correlated) habitat map}

\item{critter_correlations}{leave as NA to randomly simulate correlations across species. Otherwise, a n_species x n_species correlation matrix}

\item{resolution}{the resolution of the system}

\item{patch_area}{}
}
\value{
a list with simulated critter habitats
}
\description{
Simulate species habitats with correlations across space and among species
}
\examples{

n_species <- 6

critters <- paste0(fruit[1:n_species],"_fish")

resolution <- c(4,20)

patch_area =  4

kp = .1


species_distributions <- sim_recruitment_deviations(critters = critters, resolution = resolution, patch_area = patch_area, kp = kp)
}
