% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ref_points.R
\name{assign_ref_points}
\alias{assign_ref_points}
\title{Assign References Points}
\usage{
assign_ref_points(fauna, fleets)
}
\arguments{
\item{fauna}{a list of critters}

\item{fleets}{a list of cleets}
}
\value{
a fauna object but with MSY based reference points included for each critter
}
\description{
Assign References Points
}
