% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_class.R
\name{bird}
\alias{bird}
\title{Title}
\description{
Title

Title
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Bird-new}{\code{bird$new()}}
\item \href{#method-Bird-set_hair}{\code{bird$set_hair()}}
\item \href{#method-Bird-greet}{\code{bird$greet()}}
\item \href{#method-Bird-clone}{\code{bird$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bird-new"></a>}}
\if{latex}{\out{\hypertarget{method-Bird-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bird$new(name = NA, hair = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bird-set_hair"></a>}}
\if{latex}{\out{\hypertarget{method-Bird-set_hair}{}}}
\subsection{Method \code{set_hair()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bird$set_hair(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bird-greet"></a>}}
\if{latex}{\out{\hypertarget{method-Bird-greet}{}}}
\subsection{Method \code{greet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bird$greet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bird-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Bird-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bird$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
