% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_steps.R
\name{clean_steps}
\alias{clean_steps}
\title{A small function to trim any characters from step names in the output of \code{marlin::simmar}}
\usage{
clean_steps(step)
}
\arguments{
\item{step}{The name of the step in question}
}
\value{
the step in units of {year}_{season index}
}
\description{
A small function to trim any characters from step names in the output of \code{marlin::simmar}
}
\examples{
clean_steps("step_1_2")

}
