% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-critter.R, R/create_critter.R
\name{create_critter}
\alias{create_critter}
\title{Create Critter}
\usage{
create_critter(
  common_name = "white seabass",
  scientific_name = NA,
  critter_type = "fish",
  seasonal_habitat = list(),
  habitat_seasons = list(seasons),
  rec_habitat = NA,
  seasons = 1,
  rec_form = 1,
  adult_movement = 0,
  adult_movement_sigma = 2,
  fished_depletion = 0.4,
  resolution = 20,
  ...
)

create_critter(
  common_name = "white seabass",
  scientific_name = NA,
  critter_type = "fish",
  seasonal_habitat = list(),
  habitat_seasons = list(seasons),
  rec_habitat = NA,
  seasons = 1,
  rec_form = 1,
  adult_movement = 0,
  adult_movement_sigma = 2,
  fished_depletion = 0.4,
  resolution = 20,
  ...
)
}
\description{
Creates a critter object. If only a scientific name is provided
create_critter will try and look up relevant life history from
FishLife

Creates a critter object. If only a scientific name is provided
create_critter will try and look up relevant life history from
FishLife
}
\details{
Critical inputs are adult_movement, adult_movement_sigma, and resolution

Critical inputs are adult_movement, adult_movement_sigma, and resolution
}
