% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_critter.R
\name{create_critter}
\alias{create_critter}
\title{create_fish}
\usage{
create_critter(
  common_name = "white seabass",
  scientific_name = "Atractoscion nobilis",
  linf = NA,
  vbk = NA,
  t0 = -0.1,
  cv_len = 0.1,
  length_units = "cm",
  min_age = 0,
  max_age = NA,
  time_step = 1,
  weight_a = NA,
  weight_b = NA,
  weight_units = "kg",
  length_50_mature = NA,
  length_95_mature = NA,
  delta_mature = 0.1,
  age_50_mature = NA,
  age_95_mature = NA,
  age_mature = NA,
  length_mature = NA,
  m = NA,
  steepness = 0.8,
  r0 = 10000,
  density_dependence_form = 1,
  adult_movement = 2,
  larval_movement = 2,
  query_fishlife = T,
  price = 1,
  price_cv = 0,
  price_ac = 0,
  price_slope = 0,
  sigma_r = 0,
  rec_ac = 0,
  cores = 4,
  mat_mode = "age",
  default_wb = 2.8,
  tune_weight = FALSE,
  density_movement_modifier = 1,
  linf_buffer = 1.2
)
}
\arguments{
\item{price_cv}{}
}
\value{
a fish list object
}
\description{
creates a fish list object with all the life history goodies
}
\examples{
\dontrun{
white_seabass = create_critter(scientific_name = "Atractoscion nobilis", query_fishlife = T)
}
}
