% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_critter.R
\name{create_critter}
\alias{create_critter}
\title{Create Critter}
\usage{
create_critter(
  common_name = "white seabass",
  scientific_name = NA,
  linf = NA,
  vbk = NA,
  t0 = -0.1,
  cv_len = 0.1,
  length_units = "cm",
  min_age = 0,
  max_age = NA,
  time_step = 1,
  weight_a = NA,
  weight_b = NA,
  weight_units = "kg",
  length_50_mature = NA,
  length_95_mature = NA,
  delta_mature = 0.1,
  age_50_mature = NA,
  age_95_mature = NA,
  age_mature = NA,
  length_mature = NA,
  m = NA,
  steepness = 0.8,
  r0 = 10000,
  ssb0 = NA,
  density_dependence_form = 1,
  adult_movement = 2,
  adult_movement_sigma = 2,
  larval_movement = 2,
  query_fishlife = T,
  sigma_r = 0,
  rec_ac = 0,
  cores = 4,
  mat_mode = "age",
  default_wb = 2.8,
  tune_weight = FALSE,
  density_movement_modifier = 1,
  linf_buffer = 1.2,
  resolution = 25,
  habitat = NA,
  fished_depletion = 1,
  rec_form = 1,
  burn_steps = 100,
  seasonal_hab = NA
)
}
\arguments{
\item{common_name}{}

\item{scientific_name}{}

\item{linf}{}

\item{vbk}{}

\item{t0}{}

\item{cv_len}{}

\item{length_units}{}

\item{min_age}{}

\item{max_age}{}

\item{time_step}{}

\item{weight_a}{}

\item{weight_b}{}

\item{weight_units}{}

\item{length_50_mature}{}

\item{length_95_mature}{}

\item{delta_mature}{}

\item{age_50_mature}{}

\item{age_95_mature}{}

\item{age_mature}{}

\item{length_mature}{}

\item{m}{}

\item{steepness}{}

\item{r0}{}

\item{ssb0}{}

\item{density_dependence_form}{}

\item{adult_movement}{}

\item{adult_movement_sigma}{}

\item{larval_movement}{}

\item{query_fishlife}{}

\item{sigma_r}{}

\item{rec_ac}{}

\item{cores}{}

\item{mat_mode}{}

\item{default_wb}{}

\item{tune_weight}{}

\item{density_movement_modifier}{}

\item{linf_buffer}{}

\item{resolution}{}

\item{habitat}{}

\item{fished_depletion}{}

\item{rec_form}{}

\item{burn_steps}{}
}
\value{

}
\description{
Creates a critter object. If only a scientific name is provided
create_critter will try and look up relevant life history from
FishLife
}
\details{
Critical inputs are adult_movement, adult_movement_sigma, and resolution
}
