% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_critter.R
\name{create_critter}
\alias{create_critter}
\title{Create Critter}
\usage{
create_critter(
  common_name = NA,
  scientific_name = NA,
  get_common_name = FALSE,
  critter_type = "fish",
  base_habitat = list(),
  season_blocks = list(),
  recruit_habitat = NA,
  seasons = 1,
  fec_form = "weight",
  adult_diffusion = 2,
  recruit_diffusion = 10,
  fished_depletion = 0.4,
  init_explt = 0.1,
  explt_type = "f",
  burn_years = 50,
  weight_a = NA,
  fec_expo = 1,
  resolution = 20,
  spawning_seasons = NA,
  density_dependence = "global_habitat",
  ...
)
}
\arguments{
\item{common_name}{}

\item{scientific_name}{}

\item{critter_type}{}

\item{base_habitat}{}

\item{season_blocks}{}

\item{seasons}{}

\item{fished_depletion}{}

\item{init_explt}{initial annual exploitation rate}

\item{explt_type}{f or fmsy}

\item{...}{}

\item{rec_habitat}{}

\item{adult_movement}{}

\item{adult_movement_sigma}{}
}
\description{
Creates a critter object. If only a scientific name is provided
create_critter will try and look up relevant life history from
FishLife
}
\details{
Critical inputs are adult_movement, adult_movement_sigma, and resolution
}
