% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_critter.R
\name{create_critter}
\alias{create_critter}
\title{Create Critter}
\usage{
create_critter(
  common_name = NA,
  scientific_name = NA,
  get_common_name = FALSE,
  critter_type = "fish",
  habitat = list(),
  season_blocks = list(),
  recruit_habitat = NA,
  seasons = 1,
  fec_form = "weight",
  adult_diffusion = 2,
  recruit_diffusion = 10,
  fished_depletion = 0.4,
  init_explt = 0.1,
  explt_type = "f",
  burn_years = 50,
  weight_a = NA,
  fec_expo = 1,
  resolution = c(10, 10),
  patch_area = 1,
  spawning_seasons = NA,
  density_dependence = "global_habitat",
  ...
)
}
\arguments{
\item{common_name}{the common name of the species}

\item{scientific_name}{the scientific name of the species, preferable to common name}

\item{habitat}{a list with adult habitat per season}

\item{season_blocks}{a list with seasons per block}

\item{recruit_habitat}{habitat for recruitment}

\item{seasons}{number of seasons per year (integer)}

\item{fec_form}{one of "weight" or "pups"}

\item{adult_diffusion}{adult diffusion rate}

\item{recruit_diffusion}{recruit diffusion rate}

\item{fished_depletion}{depletion (biomass / unfished biomass) at start of simulation}

\item{init_explt}{initial annual exploitation rate (fraction of exploitable population killed)}

\item{explt_type}{f or fmsy}

\item{burn_years}{number of years to burn in the simulation prior to starting things}

\item{fec_expo}{exponent for fecundity relationship. >1 means hyperallometric fecundity}

\item{resolution}{the resolution of the system, either an integer or a vector integers of length two specifying the dimensions of the system in width and height (e.g. \code{c(10,100)})}

\item{patch_area}{the area of each patch}

\item{spawning_seasons}{which seasons spawning occurs}

\item{density_dependence}{one of 'global_habitat','local_habitat','pre_dispersal','post_dispersal','global_ssb'}
}
\description{
Creates a critter object. If only a scientific name is provided
create_critter will try and look up relevant life history from
FishLife
}
\details{
Critical inputs are adult_movement, adult_movement_sigma, and resolution
}
