% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fleet.R
\name{create_fleet}
\alias{create_fleet}
\title{create_fleet}
\usage{
create_fleet(
  fleet_model = "constant-effort",
  effort_allocation = "gravity",
  initial_effort = 100,
  sigma_effort = 0,
  effort_ac = 0
)
}
\arguments{
\item{effort_ac}{}

\item{delta}{cm above length50 at 95 selectivity}
}
\value{
a fleet object
}
\description{
create_fleet
}
\examples{
\dontrun{
create_fleet(eq_f = 2,length_50_sel = 25, length_95_sel = 27, fish = bluefish)
}
}
