% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fleet.R
\name{create_fleet}
\alias{create_fleet}
\title{Create Fleet}
\usage{
create_fleet(
  metiers,
  mpa_response = "stay",
  fleet_model = "constant_effort",
  responsiveness = 0.5,
  cost_per_unit_effort = 1,
  spatial_allocation = "rpue",
  effort_cost_exponent = 1,
  ports = NULL,
  cost_per_distance = 1,
  cr_ratio = 1,
  resolution,
  patch_area = 1,
  base_effort = NULL,
  fishing_grounds = NULL
)
}
\arguments{
\item{metiers}{a list of metiers}

\item{mpa_response}{one of "stay" or "leave" indicating response of vessels that used to fish in MPA to MPA}

\item{fleet_model}{which fleet model to use, one of "constant_effort" or "open_access" or constant catch}

\item{responsiveness}{how responsive the fleet is to profits when fleet_model = "open_access"}

\item{cost_per_unit_effort}{the cost per unit effort}

\item{spatial_allocation}{spatial effort allocation strategy ('revenue','rpue','profit','ppue')}

\item{effort_cost_exponent}{exponent of costs}

\item{ports}{location of fishing ports}

\item{cost_per_distance}{cost per unit distance}

\item{cr_ratio}{cost to revenue ratio at initial conditions (1 implies OA equilibrium, total profits = 0)}

\item{resolution}{spatial resolution of the simulated seascape}

\item{patch_area}{the area of each patch (KM^2^)}

\item{base_effort}{base effort for the fleet}

\item{fishing_grounds}{the location of fishing grounds (TRUE or FALSE)}
}
\value{
a fleet object
}
\description{
Creates a fleet object, mostly by adding in
selectivity at age for each fleet and species
}
