% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fleet.R
\name{create_fleet}
\alias{create_fleet}
\title{Create Fleet}
\usage{
create_fleet(fleets, fauna, base_effort = 0)
}
\arguments{
\item{fleets}{}

\item{fauna}{}

\item{base_effort}{}
}
\value{

}
\description{
Creates a fleet object, mostly by adding in
selectivity at age for each fleet and species
}
