% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-fleet.R
\name{create_fleet}
\alias{create_fleet}
\title{Create Fleet}
\usage{
create_fleet(metiers, base_effort = 0, mpa_response = "stay")
}
\arguments{
\item{base_effort}{}

\item{fleets}{}

\item{fauna}{}
}
\value{

}
\description{
Creates a fleet object, mostly by adding in
selectivity at age for each fleet and species
}
