% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-fleet.R
\name{create_fleet}
\alias{create_fleet}
\title{Create Fleet}
\usage{
create_fleet(
  metiers,
  base_effort = 0,
  mpa_response = "stay",
  fleet_model = "constant effort",
  profit_sensitivity = 0.1,
  cost_per_unit_effort = NA,
  spatial_allocation = "rpue",
  effort_cost_exponent = 1.3,
  ports = NULL,
  cost_per_distance = 1,
  cr_ratio = 1,
  resolution
)
}
\arguments{
\item{metiers}{a list of metiers}

\item{base_effort}{base effort for the fleet}

\item{mpa_response}{one of "stay" or "leave" indicating response of vessels that used to fish in MPA to MPA}

\item{fleet_model}{which fleet model to use, one of "constant effort" or "open access"}

\item{profit_sensitivity}{the profit sensitivity of the open access model}

\item{spatial_allocation}{spatial effort allocation strategy (ideal_free or revenue)}

\item{cr_ratio}{cost to revenue ratio at initial conditions (1 implies OA equilibrium, total profits = 0)}

\item{cost_per_unit_effortt}{the cost per unit effort in the open access model}
}
\value{
a fleet object
}
\description{
Creates a fleet object, mostly by adding in
selectivity at age for each fleet and species
}
