% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-msy.R
\name{find_msy}
\alias{find_msy}
\title{Find MSY
estimates MSY for an individual critter}
\usage{
find_msy(effort_mult, fauna, fleets, opt = TRUE, target_critter)
}
\arguments{
\item{effort_mult}{the multiplier on the base effort in fleet (assumes fleet model is already tuned)}

\item{fauna}{a list of critters}

\item{fleets}{a list of fleets}

\item{opt}{TRUE = optimize, FALSE = return MSY conditions}

\item{target_critter}{the name of the critter to find MSY for}
}
\value{
MSY conditions
}
\description{
Note that this is mostly an internal function to \code{assign_ref_points}
}
