% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tune_costs.R
\name{fine_tune_costs}
\alias{fine_tune_costs}
\title{Internal function to fine-tune cost parameters}
\usage{
fine_tune_costs(log_cost_mult, target, fauna, fleets, years = 25, tune_type)
}
\arguments{
\item{log_cost_mult}{log cost per unit effort multiplier}

\item{target}{the target to be achieved}

\item{fauna}{a list of fauna}

\item{fleets}{a list of fleets}

\item{years}{the number of years to run the optimization}

\item{tune_type}{what is being tuned for}
}
\value{
fine-tuned costs
}
\description{
Internal function to fine-tune cost parameters
}
