% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_tuner.R
\name{fleet_tuner}
\alias{fleet_tuner}
\title{fleet tuner}
\usage{
fleet_tuner(qs, fauna, fleets, years = 50)
}
\arguments{
\item{qs}{vector of catchability coefficients}

\item{fauna}{fauna object}

\item{fleets}{fleet object}

\item{years}{number of years to tune}
}
\value{
objective function of fleet tuner
}
\description{
finds catchability (q) by fleet such that
target fished depletion is achieved
}
