% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_tuner.R
\name{fleet_tuner}
\alias{fleet_tuner}
\title{fleet tuner}
\usage{
fleet_tuner(qs, fauna, fleets, steps = 50)
}
\arguments{
\item{qs}{}

\item{fauna}{}

\item{fleets}{}

\item{steps}{}
}
\value{

}
\description{
finds catchability (q) by fleet such that
target fished depletion is achieved
}
