% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_length_at_age.R
\name{generate_length_at_age_key}
\alias{generate_length_at_age_key}
\title{generate_length_to_age_key}
\usage{
generate_length_at_age_key(
  min_age,
  max_age,
  cv,
  k,
  linf,
  t0,
  time_step = 1,
  linf_buffer = 10
)
}
\arguments{
\item{max_age}{}
}
\value{
a length-at-age key
}
\description{
produces an age by length bins matrix with probability of being in length bin at age
}
