% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_length_at_age.R
\name{generate_length_at_age_key}
\alias{generate_length_at_age_key}
\title{generate_length_to_age_key}
\usage{
generate_length_at_age_key(
  min_age,
  max_age,
  cv,
  k,
  linf,
  t0,
  time_step = 1,
  linf_buffer = 10
)
}
\arguments{
\item{min_age}{minimum age tracked in the model. Best to leave at 0, as the model does not explicitly track delays for recruitment}

\item{max_age}{maximum age tracked by the model (individuals this age or older are tracked in the plus group)}

\item{cv}{the coefficient of variation of the length-at-age relationship (log-space)}

\item{k}{the}

\item{linf}{asymptotic length of the species in a von Bertalanffy growth function}

\item{t0}{hypothetical age at which the fish would have length 0 (e.g. -0.5 years)}

\item{time_step}{the time step the model is running on (1 / seasons)}

\item{linf_buffer}{multiplier around linf to create length at age key, taking into account that some fish will be larger than Linf}
}
\value{
a length-at-age key
}
\description{
produces an age by length bins matrix with probability of being in length bin at age
}
