% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marlin_pal.R
\name{marlin_pal}
\alias{marlin_pal}
\title{marlin pal palette}
\usage{
marlin_pal(palette = "fish_scales", reverse = FALSE, ...)
}
\arguments{
\item{palette}{}

\item{reverse}{}

\item{...}{}
}
\value{
color palletes
}
\description{
marlin pal palette
}
\examples{
  image(1:11, 9:10, as.matrix(1:11), 
  col=marlin_pal("fish_scales")(11),
  xlim = c(-4.5,12), ylim=c(0,10),
  xlab="", ylab = "", xaxt = "n", yaxt = "n", bty = "n")
  text(-1.8, 9.5, 'marlin_pal("fish_scales")', cex = 0.7)
image(1:7, 7.5:8.5, as.matrix(1:7), 
     col=marlin_pal("diverging_fish")(7),
     xlim = c(-4.5,12), ylim=c(0,10),
     xlab="", ylab = "", xaxt = "n", yaxt = "n", bty = "n", add = TRUE)
text(-1.8, 8, 'marlin_pal("diverging_fish")', cex = 0.7)
image(1:11, 6:7, as.matrix(1:11), 
     col=marlin_pal("lateral_lines")(11),
     xlim = c(-4.5,12), ylim=c(0,10),
     xlab="", ylab = "", xaxt = "n", yaxt = "n", bty = "n", add = TRUE)
text(-1.8, 6.5, 'marlin_pal("lateral_lines")', cex = 0.7)
image(1:11, 4.5:5.5, as.matrix(1:11), 
     col=marlin_pal("dark_blues")(11),
     xlim = c(-4.5,9), ylim=c(0,10),
     xlab="", ylab = "", xaxt = "n", yaxt = "n", bty = "n", add = TRUE)
text(-1.8, 5, 'marlin_pal("dark_blues")', cex = 0.7)
image(1:11, 3:4, as.matrix(1:11), 
     col=marlin_pal("sea_blues")(11),
     xlim = c(-4.5,9), ylim=c(0,10),
     xlab="", ylab = "", xaxt = "n", yaxt = "n", bty = "n", add = TRUE)
text(-1.8, 3.5, 'marlin_pal("sea_blues")', cex = 0.7)
image(1:11, 1.5:2.5, as.matrix(1:11), 
     col=marlin_pal("sky_blues")(11),
     xlim = c(-4.5,9), ylim=c(0,10),
     xlab="", ylab = "", xaxt = "n", yaxt = "n", bty = "n", add = TRUE)
text(-1.8, 2, 'marlin_pal("sky_blues")', cex = 0.7)
image(1:11, 0:1, as.matrix(1:11), 
     col=marlin_pal("sands")(11),
     xlim = c(-4.5,9), ylim=c(0,10),
     xlab="", ylab = "", xaxt = "n", yaxt = "n", bty = "n", add = TRUE)
text(-1.8, 0.5, 'marlin_pal("sands")', cex = 0.7)

}
