% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metier-class.R
\name{Metier}
\alias{Metier}
\title{R6 Class Representing a fishing fleet}
\description{
A fleet object has all the required characteristics of a fishing fleet
}
\details{
creates fleet object with spaces for selectivity, mpa response, etc.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Metier$new()}}
\item \href{#method-clone}{\code{Metier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metier$new(
  critter = NA,
  price = 10,
  sel_form = "logistic",
  sel_unit = "p_of_mat",
  sel_start = 1,
  sel_delta = 0.1,
  catchability = 0.01,
  spatial_catchability = NA,
  p_explt = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
