% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_mpa.R
\name{optimize_mpa}
\alias{optimize_mpa}
\title{Optimize MPA network}
\usage{
optimize_mpa(
  fauna,
  fleets,
  starting_conditions = NA,
  alpha = 0.33,
  max_prop_mpa = 1,
  resolution,
  prop_sampled = 0.2,
  max_delta = 2,
  workers = 6,
  bio_objective = "max_ssb",
  econ_objective = "yield",
  work_backwards = TRUE,
  patches_at_a_time = 1
)
}
\arguments{
\item{fauna}{fauna object}

\item{fleets}{fleet object}

\item{starting_conditions}{starting conditions of the simmar object}

\item{alpha}{the weight given to conservation relative to economics (range 0 to 1)}

\item{max_prop_mpa}{max proportion of cells to place in MPA}

\item{resolution}{the resolution of the simulated system}

\item{prop_sampled}{the proportion of cells to sample in each SIR iteration}

\item{workers}{number of workers for parallel process}

\item{objective}{one of max_ssb to maximize spawning stock biomass or min_loss to prioritize not having any losses}
}
\value{
a list with results of MPA optimization
}
\description{
Optimize MPA network
}
