% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-marlin.R
\name{plot_marlin}
\alias{plot_marlin}
\title{Plot Marlin
generates time series, spatial, or length composition plots from marlin run}
\usage{
plot_marlin(
  ...,
  steps_to_plot = NA,
  plot_var = "ssb",
  plot_type = "time",
  fauna = NULL,
  drop_recruits = TRUE
)
}
\arguments{
\item{...}{named or unnamed outputs of process_marlin}

\item{steps_to_plot}{which steps to include in the plot}

\item{plot_var}{the primary plotting variable, one of n,b,ssb}

\item{plot_type}{the plot type, one of time, length, or space}

\item{fauna}{a fauna object corresponding to the critters in process_marlin outputs}

\item{drop_recruits}{TRUE to drop recruits from length composition plots}
}
\value{
a plot
}
\description{
Plot Marlin
generates time series, spatial, or length composition plots from marlin run
}
\examples{
\dontrun{

plot_marlin(
`MPA: Sharks Offshore` = proc_offshore_mpa,
`No MPA` = proc_nearshore,
`MPA: Sharks Nearshore` = proc_nearshore_mpa,
)

}

}
