% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_movement.R
\name{prep_movement}
\alias{prep_movement}
\title{Prepare movement matrix}
\usage{
prep_movement(multiplier, resolution)
}
\arguments{
\item{multiplier}{multiplier for adjacency matrix}

\item{resolution}{spatial resolution}

\item{time_step}{time step in question}
}
\value{
a prepared movement matrix
}
\description{
Prepare movement matrix
}
