% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_marlin.R
\name{process_marlin}
\alias{process_marlin}
\title{Process Marlin}
\usage{
process_marlin(sim, steps_to_keep = NA, time_step = NA, keep_age = TRUE)
}
\arguments{
\item{sim}{the output from simmar}

\item{steps_to_keep}{which steps you'd like to keep}

\item{time_step}{the time step interval, as fractions of a year}

\item{keep_age}{TRUE keeps age structure, FALSE adds up across ages}
}
\value{
a tidy dataframe of population results
}
\description{
tidy arrays of outputs from marlin::simmar
}
\examples{
\dontrun{

processed_marlin <- process_marlin(sim, steps_to_keep = 100, time_step = 0.25)
}

}
