% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_step.R
\name{process_step}
\alias{process_step}
\title{Process Step
Convert step_YEAR_SEASON notation into YEAR_SEASON, YEAR, and SEASON}
\usage{
process_step(step)
}
\arguments{
\item{step}{}
}
\value{
a list with year_season, year, and season
}
\description{
Process Step
Convert step_YEAR_SEASON notation into YEAR_SEASON, YEAR, and SEASON
}
