% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota_finder.R
\name{quota_finder}
\alias{quota_finder}
\title{Title}
\usage{
quota_finder(
  fmult,
  quota,
  fauna,
  current_season,
  movement,
  f_p_a,
  last_n_p_a,
  f_p_a_fl,
  f
)
}
\arguments{
\item{fmult}{multiplier on fishing mortality}

\item{quota}{target quota}

\item{fauna}{fauna object}

\item{current_season}{current season}

\item{movement}{movement matrix}

\item{f_p_a}{fishing mortality rate at patch and age}

\item{last_n_p_a}{last numbers at patch and age}

\item{f_p_a_fl}{fishing mortality rate by patch age and fleet}
}
\value{
sum of squares of catch relative to quota
}
\description{
Title
}
