% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmar.R
\name{simmar}
\alias{simmar}
\title{simmar is the wrapper function for the marlin package}
\usage{
simmar(
  fauna = list(),
  fleets = list(),
  manager = list(),
  habitat = list(),
  years = 100,
  initial_conditions = NA,
  starting_step = 0,
  keep_starting_step = TRUE
)
}
\arguments{
\item{steps}{}
}
\description{
when passed fauna and fleet objects, simmar will advance
the population for a number of steps
}
