% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmar.R
\name{simmar}
\alias{simmar}
\title{simmar is the wrapper function for the marlin package}
\usage{
simmar(
  fauna = list(),
  fleets = list(),
  mpas = list(),
  years = 100,
  tune_unfished = 0
)
}
\arguments{
\item{fauna}{}

\item{fleets}{}

\item{mpas}{}

\item{tune_unfished}{}

\item{steps}{}
}
\value{

}
\description{
when passed fauna and fleet objects, simmar will advance
the population for a number of steps
}
