% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmar.R
\name{simmar}
\alias{simmar}
\title{simmar is the wrapper function for the marlin package}
\usage{
simmar(
  fauna = list(),
  fleets = list(),
  manager = list(),
  habitat = list(),
  years = 100,
  steps = NA,
  starting_season = NA,
  initial_conditions = NA,
  starting_step = NA,
  keep_starting_step = TRUE,
  log_rec_devs = NULL,
  cor_rec = diag(length(fauna))
)
}
\arguments{
\item{fauna}{a list of fauna objects}

\item{fleets}{a list of fleet objects}

\item{manager}{a list of management actions}

\item{habitat}{a list of habitat over time}

\item{years}{the number of years to run the simulation}

\item{initial_conditions}{initial conditions for the simulation, in the form simmar()[\link{final step}]}

\item{starting_step}{the step to start the simulation from, used to keep track of steps across multiple runs of simmar}

\item{keep_starting_step}{should the starting step by kept (TRUE) or dropped (FALSE)}
}
\value{
a list containing the results of the simulation
}
\description{
when passed fauna and fleet objects, simmar will advance
the population for a number of steps
}
