% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_marlin.R
\name{theme_marlin}
\alias{theme_marlin}
\title{ggplot2 theme for marlin}
\usage{
theme_marlin(base_size = 14, ...)
}
\arguments{
\item{...}{}
}
\value{
nothing
}
\description{
ggplot2 theme for marlin
}
\examples{

ggplot(mtcars, aes(mpg)) +
  geom_histogram() +
  marlin::theme_marlin()

}
