% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_diffusion.R
\name{tune_diffusion}
\alias{tune_diffusion}
\title{Tune Diffusion}
\usage{
tune_diffusion(home_range, mode = "opt")
}
\arguments{
\item{home_range}{the desired home range, defined as the linear distance such that 95\% of animals travel that distance or less from their starting location}

\item{mode}{one of "opt" to optimize diffusion rate or "plot" to return diffusion matrix}
}
\value{
The tuned diffusion rate
}
\description{
Tunes the diffusion parameters to achieve a desired "home range". The home range is set by
finding the diffusion rates that results in 95\% or less of animals being within "home range" distance
of their starting patch after one year.
}
\examples{

home_range = 42
diffusion_rate <- tune_diffusion(home_range)
diffusion_rate
}
