% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-fleets.R
\name{tune_fleets}
\alias{tune_fleets}
\title{tune fleets is an internal optimizer}
\usage{
tune_fleets(fauna, fleets, years = 50)
}
\arguments{
\item{fauna}{}

\item{fleets}{}

\item{years}{}
}
\value{

}
\description{
tune fleets is an internal optimizer
}
